Limit gem import per differentiation
Loading Fork and Zorki (and thuse future gems too) into the Gemfile at
once causes Capybara's variables to be overwritten by the last gem
loaded. Meaning facebook.com will always be the home page if Forki is
loaded last. So this does some fun hacks in the Gemfile to properly load
the one we've differentiated in the environment variables.