Add queuing to scraping jobs
This adds queuing to the project. Specifically this is so that, if
multiple requests come in they don't try to run in parallel and fail
immediately because chromedriver/selenium can only handle one at a time.

To let a Zenodotus know when a request is done this adds
a `callback_url` environmental variable. If its not set, nothing happens
after the call is finished. So make sure its set unless your running the
tests.